changes for 0.2, better makeimage script

+++ makeimage.sh
@@ -1,10 +1,19 @@
-#!/bin/bash
+#!/bin/sh
+if [ $(id -u) != "0" ]; then
+    echo "You must be the superuser to run this script" >&2
+    exit 1
+fi
+
 DRIVE='/dev/sdc'
 PART1=$DRIVE'1'
 PART2=$DRIVE'2'
 
 echo "Version:"
 read IMGVERSION
+
+echo "Zero root (y/N)?"
+read ZEROROOT
+
 MNT='/mnt/tmp'$IMGVERSION
 MNT2='/mnt/tmp'$IMGVERSION'2'
 
@@ -13,7 +22,7 @@ cd /data/pi
 mkdir $MNT
 mkdir $MNT2
 mount $PART1 $MNT
-cp $MNT/config /tmp
+cp -r $MNT/config /tmp
 rm -r $MNT/config
 cp -r /data/pi/config $MNT
 #apple
@@ -29,7 +38,7 @@ rm $MNT/config/._.Trashes
 rm $MNT/config/*.DS_Store
 
 echo "Zero FAT"
-dd if=/dev/zero of=$MNT/zero bs=1M
+#dd if=/dev/zero of=$MNT/zero bs=1M
 rm $MNT/zero
 
 mount $PART2 $MNT2
@@ -44,14 +53,16 @@ rm -r $MNT2/root/.cache
 rm -r $MNT2/root/.gstreamer-0.10
 #security
 rm $MNT2/root/.bash_history
-rm -r $MNT2/root/.ssh
+#rm -r $MNT2/root/.ssh
 #old stuff
 rm -r $MNT2/boot.bk
 rm -r $MNT2/lib/modules.bk
 
-echo "Zero Root FS"
-dd if=/dev/zero of=$MNT2/zero bs=1M
-rm $MNT2/zero
+if [ "$ZEROROOT" = "Y" -o "$ZEROROOT" = "y" ]; then
+    echo "Zero Root FS"
+    dd if=/dev/zero of=$MNT2/zero bs=1M
+    rm $MNT2/zero
+fi
 
 echo "wait 10 sec for mount"
 sleep 10
@@ -60,12 +71,12 @@ umount $MNT
 umount $MNT2
 rmdir $MNT2
 
-echo "DD"
-dd bs=1M if=$DRIVE of=musicbox$IMGVERSION.img
+echo "DD 2048 * 1M"
+dd bs=1M if=$DRIVE of=musicbox$IMGVERSION.img count=2048
 
 echo "Copy Config back"
 mount $PART1 $MNT
-cp -r /tmp/config $mnt
+cp -r /tmp/config $MNT
 rm -r /tmp/config
 
 echo "wait 10 sec for mount"
@@ -73,8 +84,8 @@ sleep 10
 umount $MNT
 rmdir $MNT
 
-echo "cut image"
-dd if=/dev/zero of=musicbox$IMGVERSION.img bs=1 count=0 seek=2G
+#echo "cut image"
+#dd if=/dev/zero of=musicbox$IMGVERSION.img bs=1 count=0 seek=2G
 
 echo "zip image"
 zip -9 musicbox$IMGVERSION.zip musicbox$IMGVERSION.img
